Refactor branches panel into frame
